dfesite/requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
Brotli==1.0.9
bs4==0.0.1
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
dateparser==1.0.0
dbf-light==1.0.0
dj-database-url==0.5.0
Django==3.1.3
django-mathfilters==1.0.0
django-plotly-dash==1.4.2
django-session-timeout==0.1.0
dpd-components==0.1.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.0.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
lxml==4.6.1
MarkupSafe==1.1.1
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.0.1
plotly==4.12.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycbrf==0.4.0
pypiwin32==223
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2020.4
pywin32==300
regex==2020.11.13
requests==2.25.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
transliterate==1.10.2
tzlocal==2.1
unoconv==0.9.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==1.2.0
XlsxWriter==1.3.7
zipp==3.4.0
